+++ client/src/components/grants/Details.tsx
@@ -7,8 +7,8 @@ import { RootState } from "../../reducers";
 import { FormInputs, Metadata, Project } from "../../types";
 import ProjectDetailsHeader from "../ProjectDetailsHeader";
 import About from "./About";
-import Rounds from "./Rounds";
-import Stats from "./Stats";
+import Rounds from "./rounds/Rounds";
+import Stats from "./stats/Stats";
 
 export default function Details({
   project,
